Major refactor of main()
The main function used to do everything and then some, factoring out
file open/read, option parsing, file/line parsing and report generation
are the first steps in making it easier to read and understand. This
should also make it easier to write tests for those parts of the code.

Miscellania:
- Updates README - the test suite is compatible with 2.6 and 3.2
- Made use of Python's implicit line continuations instead of using \
- Re-flowed a few longer lines to make it more obvious what is part of
  inner function calls.
- Updated/added assorted Lint disable-lines since pylint has trouble
  with hashlib, among other things.
- Updated copyright